fix: tag-release

+++ dent/lib/Semver/TagRelease.ts
@@ -1,4 +1,4 @@
-import { SemVer } from '../../deps.ts'
+import { SemVer, exists } from '../../deps.ts'
 
 import { GIT } from '../Git.ts'
 import { DentOptions } from '../DentOptions.ts'
@@ -20,7 +20,14 @@ export async function TagRelease(args: TagReleaseOptions): Promise<void> {
   }
 
   if (args['version-file']) {
-    await Deno.writeTextFile('VERSION', nextver)
+    if (await exists('VERSION')) {
+      const existing = await Deno.readTextFile('VERSION')
+      if (existing.trim() !== nextver) {
+        await Deno.writeTextFile('VERSION', nextver)
+      }
+    } else {
+      await Deno.writeTextFile('VERSION', nextver)
+    }
   }
 
   if (args['dry-run'] !== true) {
